feat(refactor): convert to TypeScript for better code quality.
