--- frontend/src/components/sum.js
@@ -1,7 +1,0 @@
-function sum(a, b) {
-  return a + b;
-}
-function subtract(a,b) {
-  return a - b;
-}
-export {sum, subtract};
