+++ frontend/src/components/sum.ts
@@ -1,0 +1,7 @@
+function sum(a, b) {
+  return a + b;
+}
+function subtract(a,b) {
+  return a - b;
+}
+export {sum, subtract};
